relocated pages to new page folder
