+++ src/App.css
@@ -1,4 +1,6 @@
 .App {
+  display: flex;
+  flex-direction: row;
   text-align: center;
 }
 
@@ -36,3 +38,4 @@
     transform: rotate(360deg);
   }
 }
+
