+++ src/App.js
@@ -1,10 +1,10 @@
 import "./App.css";
 import SideBar from "./components/Sidebar/Sidebar";
 
-import AddMenuItemForm from "./components/AddMenuItemForm/AddMenuItemForm";
-import AddIngredientForm from "./components/AddIngredientForm/AddIngredientForm";
-import MenuItemList from "./components/MenuItemList/MenuItemList";
-import IngredientList from "./components/IngredientList/IngredientList";
+import AddMenuItemPage from "./pages/AddMenuItemPage/AddMenuItemPage";
+import AddIngredientPage from "./pages/AddIngredientPage/AddIngredientPage";
+import MenuItemListPage from "./pages/MenuItemListPage/MenuItemListPage";
+import IngredientListPage from "./pages/IngredientListPage/IngredientListPage";
 
 import { ReactDOM } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
@@ -12,16 +12,19 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 function App() {
   return (
     <div className="App">
-      <SideBar></SideBar>
+      <SideBar className="sidebar"></SideBar>
 
       <BrowserRouter>
         <Routes>
-          <Route path="/" element={<MenuItemList/>}></Route>
+          <Route path="/" element={<MenuItemListPage/>}></Route>
 
-          <Route path="add-ingredient-form" element={<AddIngredientForm/>}></Route>
-          <Route path="add-menu-item-form" element={<AddMenuItemForm/>}></Route>
-          <Route path="ingredient-list" element={<IngredientList/>}></Route>
-          <Route path="menu-item-list" element={<MenuItemList/>}></Route>
+
+          <Route path="menu-item-list" element={<MenuItemListPage/>}></Route>
+          <Route path="ingredient-list" element={<IngredientListPage/>}></Route>
+          <Route path="add-menu-item" element={<AddMenuItemPage/>}></Route>
+          <Route path="add-ingredient" element={<AddIngredientPage/>}></Route>
+
+          
 
           <Route path="*" element={<div>404 Page not found</div>} />
         </Routes>
